Use git commit messages as a message board
Git commit messages can be used as a message board, your personal diary
a blog, or a distributed chat application.

Instead of saving your text in the files and commiting them, just place
the text to the commit message. It's an append-only operation, the
existing records do not change. Several users can add messages, you will
never have conflicts.

The command-line tooling or a web interface can be created. But why to
bother if you already have git?

HOWTO

To create a post: git commit --allow-empty

To read, search and filter the posts: git log
--since, --until, --author, --grep options can be useful.

To reference previous messages use a commit's hash.
To discuss a post create a branch.
One can create a local branch and use it as a pointer to track what
messages have been read.

If you use git forum in a small team, just give write access to the repo
to the all members.

If the number of users is big, give the write access to the moderators.

This blog is placed initially on github. There is no option to give
write access to everyone on github, so I will use pull requests to accept
messages from other users.

Feel free to comment on this post and send me a merge request.

Let's go!